# deploy/endpoints/openapi.yaml
swagger: "2.0"
info:
  title: "companeon-api"
  description: "Companeon API Gateway - Cloud Endpoints in front of Cloud Run services."
  version: "1.0.0"
host: companeon-api.endpoints.companeon.cloud.goog
...
schemes:
  - https
produces:
  - application/json
consumes:
  - application/json
securityDefinitions:
  api_key:
    type: apiKey
    name: x-api-key
    in: header
security:
  - api_key: []
...
paths:
...
  /wallet/status:
    get:
      operationId: getWalletStatus
      summary: "Get wallet delegation status"
      parameters:
        - name: wallet
          in: query
          required: true
          type: string
          description: "Wallet address"
      x-google-backend:
        address: https://companeon-api-440170696844.us-central1.run.app/wallet/status
        path_translation: CONSTANT_ADDRESS
      x-google-quota:
        metricCosts:
          read-requests: 1
      responses:
        '200':
          description: "Wallet status"
  /wallet/limits:
    get:
      operationId: getWalletLimits
      summary: "Get wallet spending limits"
      parameters:
        - name: wallet
          in: query
          required: true
          type: string
          description: "Wallet address"
      x-google-backend:
        address: https://companeon-api-440170696844.us-central1.run.app/wallet/limits
        path_translation: CONSTANT_ADDRESS
      x-google-quota:
        metricCosts:
          read-requests: 1
      responses:
        '200':
          description: "Wallet limits and scopes"
...
  /prices/range:
    get:
      operationId: getHistoricalPrices
      summary: "Get historical price data"
      parameters:
        - name: token
          in: query
          required: true
          type: string
        - name: from
          in: query
          required: false
          type: string
        - name: to
          in: query
          required: false
          type: string
      x-google-backend:
        address: https://companeon-api-440170696844.us-central1.run.app/prices/range
        path_translation: CONSTANT_ADDRESS
      x-google-quota:
        metricCosts:
          read-requests: 1
      responses:
        '200':
          description: "Historical prices"

  # ============================================
  # AGENT ENDPOINTS (API Service)
  # ============================================
...
  /agent/logs:
    get:
      operationId: getAgentLogs
      summary: "Get agent execution logs"
      parameters:
        - name: wallet
          in: query
          required: true
          type: string
        - name: limit
          in: query
          required: false
          type: integer
          default: 50
      x-google-backend:
        address: https://companeon-api-440170696844.us-central1.run.app/agent/logs
        path_translation: CONSTANT_ADDRESS
      x-google-quota:
        metricCosts:
          read-requests: 1
      responses:
        '200':
          description: "Agent logs"

  # ============================================
  # LIVE AGENT ENDPOINTS (Agent Service)
  # ============================================
...